# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Academic OS Integration API",
    "description": "Thin API Integration Facade connecting the React Frontend to the Academic OS Python Learning Intelligence Subsystem.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/materials": {
      "get": {
        "tags": [
          "Course Materials"
        ],
        "summary": "List Materials",
        "description": "List all course materials uploaded for the student.",
        "operationId": "list_materials_materials_get",
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "course",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Course"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Student-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/materials/upload": {
      "post": {
        "tags": [
          "Course Materials"
        ],
        "summary": "Upload Material",
        "description": "Upload and index a course document (.txt, .md, .pdf, .docx, .pptx).\nExtracts text, generates embeddings, and indexes chunks into Chroma vector store.",
        "operationId": "upload_material_materials_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Student-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_material_materials_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/assessments/generate": {
      "post": {
        "tags": [
          "Assessments & Knowledge Checks"
        ],
        "summary": "Generate Assessment",
        "description": "Generate an assessment / diagnostic knowledge check.\n- Grounded in uploaded material or trusted external evidence.\n- Questions contain MCQs and short answers with an explicit 'I don't know' option.\n- Stored in server cache until submitted.",
        "operationId": "generate_assessment_assessments_generate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Student-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{assessment_id}/submit": {
      "post": {
        "tags": [
          "Assessments & Knowledge Checks"
        ],
        "summary": "Submit Assessment",
        "description": "Submit completed answers for evaluation.\n- Backend evaluates correctness and updates student mastery.\n- React does NOT evaluate or update mastery client-side.",
        "operationId": "submit_assessment_assessments__assessment_id__submit_post",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-Student-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Student-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssessmentGenerateRequest": {
        "properties": {
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "AssessmentGenerateRequest"
      },
      "AssessmentGenerateResponse": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/AssessmentQuestionModel"
            },
            "type": "array",
            "title": "Questions"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "topic",
          "questions",
          "source_type",
          "source_title"
        ],
        "title": "AssessmentGenerateResponse"
      },
      "AssessmentQuestionModel": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "concept": {
            "type": "string",
            "title": "Concept"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "medium"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": "trusted_external"
          },
          "source_title": {
            "type": "string",
            "title": "Source Title",
            "default": "Authoritative Sources"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "concept",
          "kind",
          "prompt"
        ],
        "title": "AssessmentQuestionModel"
      },
      "AssessmentSubmitRequest": {
        "properties": {
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "answers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Answers"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "AssessmentSubmitRequest"
      },
      "AssessmentSubmitResponse": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "concept_mastery": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Concept Mastery"
          },
          "weak_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Concepts"
          },
          "unknown_concepts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unknown Concepts"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "feedback_per_question": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Per Question"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "score",
          "concept_mastery",
          "weak_concepts",
          "unknown_concepts",
          "strengths",
          "recommendations"
        ],
        "title": "AssessmentSubmitResponse"
      },
      "Body_upload_material_materials_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "course": {
            "type": "string",
            "title": "Course",
            "default": "General"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_material_materials_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialModel": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "course": {
            "type": "string",
            "title": "Course"
          },
          "upload_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Date"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "indexed"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 1
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": "student_upload"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "course"
        ],
        "title": "MaterialModel"
      },
      "MaterialUploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "course": {
            "type": "string",
            "title": "Course"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "indexed"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "File successfully uploaded and indexed into knowledge base."
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "course",
          "chunk_count"
        ],
        "title": "MaterialUploadResponse"
      },
      "MaterialsListResponse": {
        "properties": {
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialModel"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "title": "MaterialsListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}